// lib/_TC/_Ctx/enums/BrdBlm_TC_Ctx_eInternBaseboardFinish.ts
/** ---------------------------------------------------------------------------
 * @module [BrdBlm_TC_Ctx]
 * @author [APG] Angeli Paolo Giusto 
 * @version 0.1 APG 20240323
 * ----------------------------------------------------------------------------
 */

/**
 * Texture per gli zoccoli interni del contesto (albedo e bump o normal maps)
 */
export enum BrdBlm_TC_Ctx_eInternBaseboardFinish {
    BASEBOARD_1 = 'context_indoor_baseboard_1',
    BASEBOARD_2 = 'context_indoor_baseboard_2',
    BASEBOARD_3 = 'context_indoor_baseboard_3',
}

/*! ---------------------------------------------------------------------------
 * @copyright Breda Sistemi industriali S.p.A., 2023 - http://bredasys.com
 * All rights reserved
 * @licence You cannot host, display, distribute or share this Work in any
 * form, both physical and digital. You cannot use this Work in any commercial
 * or non-commercial product, website or project. You cannot sell this Work
 * and you cannot mint an NFTs out of it.
 * ---------------------------------------------------------------------------
 */